Remote refactor
